Ensure security demo initializes asynchronously
