+++ index.html
@@ -228,6 +228,6 @@
     </div>
   </div>
 
-  <script src="script.js" defer></script>
+  <script type="module" src="main.js"></script>
 </body>
 </html>
