+++ theme.js
@@ -1,0 +1,68 @@
+export function initTheme() {
+  const themeToggle = document.querySelector('.theme-toggle');
+  const body = document.body;
+  if (!themeToggle) return;
+  const icon = themeToggle.querySelector('i');
+  const currentTheme = localStorage.getItem('theme');
+  if (currentTheme === 'dark') {
+    body.classList.add('dark-mode');
+    icon.classList.replace('fa-moon', 'fa-sun');
+  }
+
+  themeToggle.addEventListener('click', () => {
+    body.classList.toggle('dark-mode');
+
+    if (body.classList.contains('dark-mode')) {
+      icon.classList.replace('fa-moon', 'fa-sun');
+      localStorage.setItem('theme', 'dark');
+    } else {
+      icon.classList.replace('fa-sun', 'fa-moon');
+      localStorage.setItem('theme', 'light');
+    }
+  });
+
+  const navbar = document.querySelector('.navbar');
+  window.addEventListener('scroll', () => {
+    if (window.scrollY > 100) {
+      navbar.style.background = body.classList.contains('dark-mode')
+        ? 'rgba(18, 18, 18, 0.98)'
+        : 'rgba(255, 255, 255, 0.98)';
+      navbar.style.boxShadow = '0 4px 20px rgba(0, 0, 0, 0.1)';
+    } else {
+      navbar.style.background = body.classList.contains('dark-mode')
+        ? 'rgba(18, 18, 18, 0.95)'
+        : 'rgba(255, 255, 255, 0.95)';
+      navbar.style.boxShadow = '0 2px 10px rgba(0, 0, 0, 0.1)';
+    }
+  });
+}
+
+export function initNavigation() {
+  const navToggle = document.querySelector('.nav-toggle');
+  const navMenu = document.querySelector('.nav-menu');
+  if (!navToggle || !navMenu) return navMenu;
+
+  navToggle.addEventListener('click', () => {
+    navMenu.classList.toggle('active');
+    const spans = navToggle.querySelectorAll('span');
+    spans[0].style.transform = navMenu.classList.contains('active')
+      ? 'rotate(-45deg) translate(-5px, 6px)'
+      : '';
+    spans[1].style.opacity = navMenu.classList.contains('active') ? '0' : '1';
+    spans[2].style.transform = navMenu.classList.contains('active')
+      ? 'rotate(45deg) translate(-5px, -6px)'
+      : '';
+  });
+
+  document.querySelectorAll('.nav-menu a').forEach((link) => {
+    link.addEventListener('click', () => {
+      navMenu.classList.remove('active');
+      const spans = navToggle.querySelectorAll('span');
+      spans[0].style.transform = '';
+      spans[1].style.opacity = '1';
+      spans[2].style.transform = '';
+    });
+  });
+
+  return navMenu;
+}
